0.4 MK3
Added new Claudio
Fixed Highmax super
added new highmax moves
Added new teching system for regular attacks
fixed vile gizmo grab not dealing flinch
Added new zain gab interaction
Added new highmax stunshot
Buffed storm eagle dive attack
Added new armadillo and kuwanger moves from wcut 95
Added new Single Player Door System
Added new single player X triggers

+++ src/Weapons/VileCutter.cs
@@ -251,9 +251,9 @@ public class VileMaroonedTomahawk : Projectile {
 			stopMoving();
 		}
 
-		if (owner.input.isPressed(Control.Taunt, owner)) {
+		if (owner.input.isPressed(Control.Taunt, owner) && ownerActor != null) {
 			new TwoHeadedSlash(
-				pos, xDir, owner.character, owner,
+				pos, xDir, ownerActor, owner,
 				owner.getNextActorNetId(), rpc: true);
 			destroySelf();
         }
@@ -426,7 +426,7 @@ public class MetalCrescent : Projectile {
 		Point pos, int xDir,
 		Actor owner, Player player, ushort? netId, bool rpc = false
 	) : base(
-		pos, xDir, owner, "cutter_qh", netId, player
+		pos, xDir, owner, "cutter_mc", netId, player
 	) {
 	
 		damager.damage = 2;
